Converted Population object to operate on templatized Organism, FitnessEvaluator and Mutator classes.
Removed analysis methods from Population and put them in a new class,
GenebankAnalyzer, which is templatized on Organism (no need for it to
know about FitnessEvaluator and Mutator).

Implemented two new classes (SimpleMutator and Polymerase) to serve as
Mutators in the Population class.  SimpleMutator implements the
previous mutation process, while Polymerase will (in the future) house
a more elaborate mutational model with transition probabilities.

Updated tests and tools to follow new Population semantics.

Will add mutator.* in a second -- oops!



git-svn-id: http://wlab.ccbb.utexas.edu/svn/evoli@190 dadfaeae-b11a-0410-95f2-80a8e7ca9aa5